+++ packages/react/src/components/WhatsNew/index.ts
@@ -1,0 +1,17 @@
+/**
+ * @license
+ *
+ * Copyright IBM Corp. 2025
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+export { Bubble, BubbleHeader } from './components/Bubble';
+export {
+  Toc,
+  TocItem,
+  TocList,
+  TocSection,
+  TocSections,
+} from './components/Toc';
+export { ViewStack, View } from './components/ViewStack';