+++ packages/web-components/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.13.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.12.0...@carbon-labs/web-components@0.13.0) (2025-03-24)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.12.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.11.0...@carbon-labs/web-components@0.12.0) (2025-03-11)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.11.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.10.0...@carbon-labs/web-components@0.11.0) (2025-03-10)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.10.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.9.0...@carbon-labs/web-components@0.10.0) (2025-03-06)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.9.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.8.0...@carbon-labs/web-components@0.9.0) (2025-03-05)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
 # [0.8.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.7.0...@carbon-labs/web-components@0.8.0) (2025-03-03)
 
 **Note:** Version bump only for package @carbon-labs/web-components